Refactor to separate UI from logic
