+++ app/components/timeGroup/timeGroupMore/utils/types.tsx
@@ -1,6 +1,4 @@
 import { T_GameState } from "@/app/utils/types";
-import {  T_AllToDustOutput } from "../../../../utils/types";
-
 
 export interface I_MoreSections {
     leftHeadingWidth : string,
